requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pillow
torch
torchvision
transformers
